migrate tests to integ tests to run as separate processes
