+++ src/lib.rs
@@ -363,183 +363,4 @@ where
     }
 }
 
-// Use nextest to run these, otherwise there will be race conditions against the strategy oncecell
-#[cfg(test)]
-mod tests {
-    use std::any::Any;
-
-    use super::*;
-
-    #[cfg(not(feature = "tokio"))]
-    #[tokio::test]
-    async fn default_to_current_context_non_tokio() {
-        let future = async { 5 };
-        
-        let res = spawn_compute_heavy_future(future).await.unwrap();
-        assert_eq!(res, 5);
-
-        assert!(matches!(COMPUTE_HEAVY_FUTURE_EXECUTOR_STRATEGY.get(), Some(&ExecutorStrategy::CurrentContext(..))));
-    }
-
-    #[cfg(feature = "tokio")]
-    #[tokio::test]
-    async fn default_to_current_context_tokio_single_threaded() {
-        // this is a tokio test but we haven't enabled the tokio config flag
-
-        let future = async { 5 };
-        
-        let res = spawn_compute_heavy_future(future).await.unwrap();
-        assert_eq!(res, 5);
-
-        assert!(matches!(COMPUTE_HEAVY_FUTURE_EXECUTOR_STRATEGY.get(), Some(&ExecutorStrategy::SpawnBlocking(..))));
-    }
-
-    #[cfg(feature = "tokio")]
-    #[tokio::test(flavor = "multi_thread")]
-    async fn default_to_current_context_tokio_multi_threaded() {
-        let future = async { 5 };
-        
-        let res = spawn_compute_heavy_future(future).await.unwrap();
-        assert_eq!(res, 5);
-
-        assert!(matches!(COMPUTE_HEAVY_FUTURE_EXECUTOR_STRATEGY.get(), Some(&ExecutorStrategy::BlockInPlace(..))));
-    }
-
-    #[tokio::test]
-    async fn current_context_strategy() {
-        initialize_current_context_strategy();
-
-        let future = async { 5 };
-
-        let res = spawn_compute_heavy_future(future).await.unwrap();
-        assert_eq!(res, 5);
-
-        assert!(matches!(COMPUTE_HEAVY_FUTURE_EXECUTOR_STRATEGY.get(), Some(&ExecutorStrategy::CurrentContext(..))));
-    }
-
-    #[tokio::test]
-    async fn custom_strategy_legal_closure() {
-        let closure: CustomExecutorClosure = Box::new(|fut| {
-            Box::new(async move {
-                let res = fut.await;
-                Ok(res)
-            })
-        });
-
-        initialize_custom_executor_strategy(closure).await;
-
-        let future = async { 5 };
-        
-        let res = spawn_compute_heavy_future(future).await.unwrap();
-        assert_eq!(res, 5);
-
-        assert!(matches!(COMPUTE_HEAVY_FUTURE_EXECUTOR_STRATEGY.get(), Some(&ExecutorStrategy::CustomExecutor(..))));
-    }
-
-    #[cfg(feature = "tokio")]
-    #[tokio::test]
-    async fn custom_strategy_legal_closure_tokio_spawn() {
-        let closure: CustomExecutorClosure = Box::new(|fut| {
-            Box::new(
-                async move {
-                    let handle = tokio::task::spawn(async move { fut.await });
-                    handle.await.map_err(|err| err.into())
-                }
-            )
-        });
-        initialize_custom_executor_strategy(closure).await;
-
-        let future = async { 5 };
-        
-        let res = spawn_compute_heavy_future(future).await.unwrap();
-        assert_eq!(res, 5);
-
-        assert!(matches!(COMPUTE_HEAVY_FUTURE_EXECUTOR_STRATEGY.get(), Some(&ExecutorStrategy::CustomExecutor(..))));
-    }
-
-    #[tokio::test]
-    #[should_panic]
-    async fn custom_strategy_illegal_closure() {
-        // this closure overwrites the input type, causing a mismatch
-        let closure: CustomExecutorClosure = Box::new(|_| {
-            Box::new(async move {
-                Ok(Box::new("foo") as Box<dyn Any + Send>)
-            })
-        });
-        // this should panic due to bad closure
-        initialize_custom_executor_strategy(closure).await;
-    }
-
-    #[cfg(feature = "tokio")]
-    #[tokio::test]
-    async fn spawn_blocking_strategy() {
-        initialize_spawn_blocking_strategy();
-
-        let future = async { 5 };
-
-        let res = spawn_compute_heavy_future(future).await.unwrap();
-        assert_eq!(res, 5);
-
-        assert!(matches!(COMPUTE_HEAVY_FUTURE_EXECUTOR_STRATEGY.get(), Some(&ExecutorStrategy::SpawnBlocking(..))));
-    }
-
-    #[cfg(feature = "tokio")]
-    #[tokio::test(flavor = "multi_thread")]
-    async fn block_in_place_strategy() {
-        initialize_block_in_place_strategy();
-
-        let future = async { 5 };
-        
-        let res = spawn_compute_heavy_future(future).await.unwrap();
-        assert_eq!(res, 5);
-
-        assert!(matches!(COMPUTE_HEAVY_FUTURE_EXECUTOR_STRATEGY.get(), Some(&ExecutorStrategy::BlockInPlace(..))));
-    }
-
-    #[cfg(feature = "tokio")]
-    #[tokio::test]
-    async fn secondary_tokio_runtime_strategy() {
-        initialize_secondary_tokio_runtime_strategy();
-
-        let future = async { 5 };
-        
-        let res = spawn_compute_heavy_future(future).await.unwrap();
-        assert_eq!(res, 5);
-
-        assert!(matches!(COMPUTE_HEAVY_FUTURE_EXECUTOR_STRATEGY.get(), Some(&ExecutorStrategy::SecondaryTokioRuntime(..))));
-    }
-
-    #[cfg(feature = "tokio")]
-    #[tokio::test]
-    async fn secondary_tokio_runtime_strategy_allowed_config() {
-        initialize_secondary_tokio_runtime_strategy_and_config(Some(5), Some(50));
-
-        let future = async { 5 };
-        
-        let res = spawn_compute_heavy_future(future).await.unwrap();
-        assert_eq!(res, 5);
-
-        assert!(matches!(COMPUTE_HEAVY_FUTURE_EXECUTOR_STRATEGY.get(), Some(&ExecutorStrategy::SecondaryTokioRuntime(..))));
-    }
-
-    #[cfg(feature = "tokio")]
-    #[tokio::test]
-    #[should_panic]
-    async fn secondary_tokio_runtime_strategy_disallowed_config() {
-        initialize_secondary_tokio_runtime_strategy_and_config(Some(30), Some(50));
-
-        let future = async { 5 };
-        
-        let res = spawn_compute_heavy_future(future).await.unwrap();
-        assert_eq!(res, 5);
-
-        assert!(matches!(COMPUTE_HEAVY_FUTURE_EXECUTOR_STRATEGY.get(), Some(&ExecutorStrategy::SecondaryTokioRuntime(..))));
-    }
-
-    #[tokio::test]
-    #[should_panic]
-    async fn multiple_initializes_panic() {
-        initialize_current_context_strategy();
-        initialize_current_context_strategy();
-    }
-}
+// tests are in /tests/ to allow seperate intialization of oncelock when using default cargo test runner
